+++ packages/core-js-compat/helpers.js
@@ -8,6 +8,18 @@ function compare(a, operator, b) {
   return cmp(semver(a), operator, semver(b));
 }
 
+function filterOutStabilizedProposals(modules) {
+  const modulesSet = new Set(modules);
+
+  for (const $module of modulesSet) {
+    if ($module.startsWith('esnext.') && modulesSet.has($module.replace(/^esnext\./, 'es.'))) {
+      modulesSet.delete($module);
+    }
+  }
+
+  return [...modulesSet];
+}
+
 function intersection(list, order) {
   const set = list instanceof Set ? list : new Set(list);
   return order.filter(name => set.has(name));
@@ -22,6 +34,7 @@ function sortObjectByKey(object, fn) {
 
 module.exports = {
   compare,
+  filterOutStabilizedProposals,
   has,
   intersection,
   semver,